fix(desktop): centralize activity feed presentation
