+++ products/desktop/packages/ui/src/features/canvas/components/ActivityView.tsx
@@ -1,22 +1,7 @@
-import {
-  BellIcon,
-  ChatCircleIcon,
-  CheckIcon,
-  ChecksIcon,
-  DotsThreeIcon,
-  LinkIcon,
-  QuestionIcon,
-} from "@phosphor-icons/react";
+import { BellIcon } from "@phosphor-icons/react";
 import type { TaskActivityItem } from "@posthog/core/canvas/taskActivity";
 import {
-  Avatar,
-  AvatarFallback,
-  Badge,
   Button,
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
   Empty,
   EmptyDescription,
   EmptyHeader,
@@ -26,182 +11,24 @@ import {
   Spinner,
 } from "@posthog/quill";
 import { ANALYTICS_EVENTS } from "@posthog/shared/analytics-events";
-import type { UserBasic } from "@posthog/shared/domain-types";
 import { useOptionalAuthenticatedClient } from "@posthog/ui/features/auth/authClient";
-import { UserAvatar } from "@posthog/ui/features/auth/UserAvatar";
 import { useCurrentUser } from "@posthog/ui/features/auth/useCurrentUser";
+import { ActivityActionsMenu } from "@posthog/ui/features/canvas/components/ActivityActionsMenu";
+import { ActivityRow } from "@posthog/ui/features/canvas/components/ActivityRow";
 import { ActivityUnreadsToggle } from "@posthog/ui/features/canvas/components/ActivityUnreadsToggle";
-import { MentionText } from "@posthog/ui/features/canvas/components/MentionText";
 import { openActivityItem } from "@posthog/ui/features/canvas/components/openActivityItem";
 import { useBlockedTaskIds } from "@posthog/ui/features/canvas/hooks/useBlockedSessionCount";
 import { useLocalDayStart } from "@posthog/ui/features/canvas/hooks/useLocalDayStart";
 import { useMarkTaskActivityRead } from "@posthog/ui/features/canvas/hooks/useMarkTaskActivityRead";
 import { useTaskActivity } from "@posthog/ui/features/canvas/hooks/useTaskActivity";
 import { useActivityFilterStore } from "@posthog/ui/features/canvas/stores/activityFilterStore";
-import { copyChannelLink } from "@posthog/ui/features/canvas/utils/copyChannelLink";
-import { useCommentNavigationStore } from "@posthog/ui/features/sessions/commentNavigationStore";
 import { track } from "@posthog/ui/shell/analytics";
-import {
-  Fragment,
-  type ReactElement,
-  useCallback,
-  useEffect,
-  useMemo,
-} from "react";
+import { Fragment, useCallback, useEffect, useMemo } from "react";
 import {
   activityReadPayload,
   getUnreadActivityItems,
   groupActivityItemsByDay,
-  markLoadedReadLabel,
 } from "./activityFeed";
-import { activityMetadata } from "./activityMetadata";
-
-function AgentActivityIcon({
-  item,
-  className,
-}: {
-  item: TaskActivityItem;
-  className?: string;
-}): ReactElement {
-  switch (item.activityKind) {
-    case "completed":
-      return <CheckIcon size={13} weight="bold" className={className} />;
-    case "awaiting_input":
-      return <QuestionIcon size={12} weight="bold" className={className} />;
-    default:
-      return <ChatCircleIcon size={13} weight="bold" className={className} />;
-  }
-}
-
-export function ActivityRow({
-  item,
-  channelId,
-  onOpen,
-  onMarkRead,
-  currentUser,
-  blockedTaskIds,
-  surface = "activity",
-  onActivate,
-  isSelected = false,
-  compact = false,
-}: {
-  item: TaskActivityItem;
-  /** Backend channel id (the /website route param); null when the task is unfiled. */
-  channelId: string | null;
-  onOpen: (item: TaskActivityItem) => void;
-  onMarkRead: (item: TaskActivityItem) => void;
-  currentUser?: UserBasic | null;
-  /**
-   * Tasks whose session is waiting on you. Passed in rather than selected here:
-   * the feed renders one of these per item, and the selector behind it scans and
-   * sorts every live session on each store notification.
-   */
-  blockedTaskIds: ReadonlySet<string>;
-  surface?: "activity" | "activity_panel";
-  /** Where the row goes is the feed's business, not the row's. */
-  onActivate: (item: TaskActivityItem) => void;
-  isSelected?: boolean;
-  compact?: boolean;
-}) {
-  const isAgentActivity =
-    item.activityKind === "awaiting_input" ||
-    item.activityKind === "completed" ||
-    (item.activityKind === "message" && !item.author);
-  // An awaiting_input row records a past event, but only the live session knows
-  // whether it still needs a reply. Active waiting therefore outranks unread color.
-  const awaitsReply =
-    item.activityKind === "awaiting_input" && blockedTaskIds.has(item.taskId);
-  const agentIconClassName = awaitsReply ? "text-(--blue-11)" : undefined;
-  const agentIconWrapperClassName =
-    item.isUnread && !awaitsReply
-      ? "bg-primary text-primary-foreground"
-      : undefined;
-  const openTask = () => {
-    track(ANALYTICS_EVENTS.CHANNEL_ACTION, {
-      action_type: "open_task",
-      surface,
-      channel_id: channelId ?? undefined,
-      task_id: item.taskId,
-    });
-    onOpen(item);
-    if (item.commentId && item.commentTarget) {
-      useCommentNavigationStore
-        .getState()
-        .requestCommentFocus(item.taskId, item.commentTarget, item.commentId);
-    }
-    onActivate(item);
-  };
-
-  return (
-    <div className="group relative">
-      <Button
-        type="button"
-        onClick={openTask}
-        left
-        className={`h-auto w-full items-start text-left ${compact ? "py-1.5 pr-10" : "py-2"} ${isSelected ? "bg-fill-selected" : ""}`}
-      >
-        <span className="mt-0.5 shrink-0">
-          {isAgentActivity ? (
-            <Avatar size="xs" className={agentIconWrapperClassName}>
-              <AvatarFallback>
-                <AgentActivityIcon item={item} className={agentIconClassName} />
-              </AvatarFallback>
-            </Avatar>
-          ) : (
-            <span className="mt-1 flex shrink-0">
-              <UserAvatar user={item.author ?? currentUser} size="xs" />
-            </span>
-          )}
-        </span>
-        <span className="min-w-0 flex-1">
-          <span className="flex items-baseline gap-2">
-            <span
-              className={`truncate text-[13px] ${item.isUnread ? "font-semibold" : "font-medium"}`}
-            >
-              {item.taskTitle}
-            </span>
-            {item.isUnread && !compact && <Badge variant="info">New</Badge>}
-          </span>
-          <span className="block truncate text-muted-foreground text-xxs">
-            {activityMetadata(item, currentUser?.email)}
-          </span>
-          {item.snippet && !compact && (
-            <MentionText
-              content={item.snippet}
-              currentUserEmail={currentUser?.email}
-              className="mt-1 block whitespace-pre-wrap break-words text-xs"
-            />
-          )}
-        </span>
-      </Button>
-      {item.isUnread && (
-        <Button
-          variant="default"
-          size="icon-xs"
-          aria-label="Mark as read"
-          title="Mark as read"
-          className={`absolute opacity-0 transition-opacity group-hover:opacity-100 ${compact ? "top-2 right-2" : `top-2 ${channelId ? "right-9" : "right-2"}`}`}
-          onClick={() => onMarkRead(item)}
-        >
-          <CheckIcon size={14} />
-        </Button>
-      )}
-      {channelId && !compact && (
-        <Button
-          variant="default"
-          size="icon-xs"
-          aria-label="Copy thread link"
-          className="absolute top-2 right-2 opacity-0 transition-opacity group-hover:opacity-100"
-          onClick={() =>
-            void copyChannelLink(channelId, "activity", item.taskId)
-          }
-        >
-          <LinkIcon size={14} />
-        </Button>
-      )}
-    </div>
-  );
-}
 
 // The Activity page for the code layout: every task the viewer is involved in —
 // created, mentioned in, or messaged in — newest activity first. Rows clear as
@@ -225,30 +52,18 @@ export function ActivityView() {
   const blockedTaskIds = useBlockedTaskIds();
   const { mutate: markTasksRead, isPending: isMarkingRead } =
     useMarkTaskActivityRead();
-  const visibleItems = items;
-  const unreadItems = useMemo(
-    () => getUnreadActivityItems(visibleItems),
-    [visibleItems],
-  );
+  const unreadItems = useMemo(() => getUnreadActivityItems(items), [items]);
   const unreadsOnly = useActivityFilterStore((state) => state.unreadsOnly);
-  const shownItems = unreadsOnly ? unreadItems : visibleItems;
+  const shownItems = unreadsOnly ? unreadItems : items;
   const dayStart = useLocalDayStart();
   const shownItemGroups = useMemo(
     () => groupActivityItemsByDay(shownItems, new Date(dayStart)),
     [shownItems, dayStart],
   );
-  const visibleUnreadCount = unreadCount;
   // Opening a row is what marks it read. The server does the same when the task is
   // reached any other way, so the feed converges either way.
   const markRead = useCallback(
-    (item: TaskActivityItem) =>
-      markTasksRead([
-        {
-          task_id: item.taskId,
-          seen_before: item.activityAt,
-          ...(item.commentId ? { activity_id: item.id } : {}),
-        },
-      ]),
+    (item: TaskActivityItem) => markTasksRead(activityReadPayload([item])),
     [markTasksRead],
   );
   const markAllRead = useCallback(() => {
@@ -312,8 +127,6 @@ export function ActivityView() {
                 <ActivityRow
                   key={item.id}
                   item={item}
-                  channelId={item.channelId}
-                  onOpen={markRead}
                   onMarkRead={markRead}
                   onActivate={openActivityItem}
                   currentUser={currentUser}
@@ -339,36 +152,12 @@ export function ActivityView() {
           </div>
           <div className="flex shrink-0 items-center gap-2">
             <ActivityUnreadsToggle />
-            <DropdownMenu>
-              <DropdownMenuTrigger
-                render={
-                  <Button
-                    type="button"
-                    variant="default"
-                    size="icon-sm"
-                    aria-label="Activity actions"
-                    loading={isMarkingRead}
-                    disabled={isMarkingRead}
-                  >
-                    <DotsThreeIcon size={14} weight="bold" />
-                  </Button>
-                }
-              />
-              <DropdownMenuContent
-                align="end"
-                side="bottom"
-                sideOffset={4}
-                className="w-max"
-              >
-                <DropdownMenuItem
-                  disabled={isMarkingRead || unreadItems.length === 0}
-                  onClick={markAllRead}
-                >
-                  <ChecksIcon size={14} />
-                  {markLoadedReadLabel(unreadItems.length, visibleUnreadCount)}
-                </DropdownMenuItem>
-              </DropdownMenuContent>
-            </DropdownMenu>
+            <ActivityActionsMenu
+              loadedUnreadCount={unreadItems.length}
+              totalUnreadCount={unreadCount}
+              isMarkingRead={isMarkingRead}
+              onMarkAllRead={markAllRead}
+            />
           </div>
         </div>
         <div className="mt-4">{feed}</div>
